+++ src/python/persistence/pymodule_persistence.cpp
@@ -3,6 +3,7 @@
 #include "py_barcode.hpp"
 #include "py_barcode_summary.hpp"
 #include "py_homological_kernel.hpp"
+#include "py_linkage.hpp"
 #include "py_persistence_pair.hpp"
 #include "py_ripser.hpp"
 
@@ -19,5 +20,6 @@ namespace sb_py
     register_persistence_persistence_pair(sm);
     register_persistence_ripser(sm);
     register_persistence_homological_kernel(sm);
+    register_persistence_linkage(sm);
   }
 }